third_party/blink/web_tests/external/wpt/editing/run/insertunorderedlist_2001-last-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}<br></div>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<div>{}<br></div>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "true" but got "false"
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}</p>": execCommand("insertunorderedlist", false, "") return value
  assert_equals: expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}</p>": execCommand("inserttext", false, "abc") return value
  assert_equals: expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}</p>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p></p>" not in array ["<ul><li>abc</li></ul>", "<ul><li>abc<br></li></ul>"]
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}</p>" queryCommandState("insertunorderedlist") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}</p>" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "true" but got "false"
[FAIL] [["insertunorderedlist",""],["inserttext","abc"]] "<p>{}<br></p>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<p><ul><li>abc</li></ul></p>" not in array ["<ul><li>abc</li></ul>", "<ul><li>abc<br></li></ul>"]
[FAIL] [["insertunorderedlist",""]] "<div id=\\"a\\">a[bc</div><div id=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div id=\\"a\\"><ul><li>abc</li><li>def</li></ul></div>" not in array ["<div><ul><li id=\\"a\\">abc</li><li id=\\"b\\">def</li></ul></div>", "<ul><li id=\\"a\\">abc</li><li id=\\"b\\">def</li></ul>"]
[FAIL] [["insertunorderedlist",""]] "<div class=\\"a\\">a[bc</div><div class=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div class=\\"a\\"><ul><li>abc</li><li>def</li></ul></div>" not in array ["<div><ul><li class=\\"a\\">abc</li><li class=\\"b\\">def</li></ul></div>", "<ul><li class=\\"a\\">abc</li><li class=\\"b\\">def</li></ul>"]
[FAIL] [["insertunorderedlist",""]] "<div title=\\"a\\">a[bc</div><div title=\\"b\\">de]f</div>" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "<div title=\\"a\\"><ul><li>abc</li><li>def</li></ul></div>" not in array ["<div><ul><li title=\\"a\\">abc</li><li title=\\"b\\">def</li></ul></div>", "<ul><li title=\\"a\\">abc</li><li title=\\"b\\">def</li></ul>"]
Harness: the test ran to completion.
